refactor: 당첨결과를 구하는 로직 위치 변경
