+++ src/main/java/domain/LottoRank.java
@@ -12,6 +12,7 @@ public enum LottoRank {
 	SECOND(30000000, 5, true),
 	FIRST(2000000000, 6, false);
 
+	private static final int INITIAL_COUNT = 0;
 	private final long amount;
 	private final int matchCount;
 	private final boolean hasBonusNumber;
@@ -31,7 +32,15 @@ public enum LottoRank {
 		return FAIL;
 	}
 
-	public static boolean isFail(LottoRank rank) {
+	public static EnumMap<LottoRank, Integer> createWinningResultMap() {
+		EnumMap<LottoRank, Integer> rankMap = new EnumMap<>(LottoRank.class);
+		Arrays.stream(values())
+			.filter(value -> !isFail(value))
+			.forEach(value -> rankMap.put(value, INITIAL_COUNT));
+		return rankMap;
+	}
+
+	public static boolean isFail(final LottoRank rank) {
 		return rank == FAIL;
 	}
